Coorodonnées maintenant int (pas double) Les combats se quittent si un combat avec aucun ennemi est activé.

+++ src/Outils/Coord.java
@@ -4,15 +4,15 @@ package Outils;
  * Created by LAMG030499 on 2017-08-22.
  */
 public class Coord {
-    public double x = 0;
-    public double y = 0;
+    public int x = 0;
+    public int y = 0;
 
     public Coord(){
         x = 0;
         y = 0;
     }
 
-    public Coord(double x, double y){
+    public Coord(int x, int y){
         this.y = y;
         this.x = x;
     }
